moved hashes into new files
