+++ swyft/utils.py
@@ -1,108 +1,47 @@
-import numpy as np
-import pylab as plt
-from scipy.interpolate import griddata
+# pylint: disable=no-member, not-callable
+from warnings import warn
 
-def get_contour_levels(x, cred_level = [0.68268, 0.95450, 0.99730]):
-    x = np.sort(x)[::-1]  # Sort backwards
-    total_mass = x.sum()
-    enclosed_mass = np.cumsum(x)
-    idx = [np.argmax(enclosed_mass >= total_mass * f) for f in cred_level]
-    levels = np.array(x[idx])
-    return levels
+import torch
 
-def cont2d(ax, re, x0, z0, i, j, cmap = 'gray_r', Nmax = 1000):
-    z, p = re.posterior(x0, [i, j], Nmax = Nmax)
-    z = z.numpy()
-    levels = get_contour_levels(p)
 
-    if z0 is not None:
-        ax.axvline(z0[i], color='r', ls=':')
-        ax.axhline(z0[j], color='r', ls=':')
-    
-    N = 100*1j
-    extent = [z[:,0].min(), z[:,0].max(), z[:,1].min(), z[:,1].max()]
-    xs,ys = np.mgrid[z[:,0].min():z[:,0].max():N, z[:,1].min():z[:,1].max():N]
-    resampled = griddata(z, p, (xs, ys))
-    ax.imshow(resampled.T, extent = extent, origin='lower', cmap=cmap, aspect = 'auto')
-    ax.tricontour(z[:,0], z[:,1], -p, levels = -levels, colors = 'k', linestyles=['-'])
-    
-def hist1d(ax, re, x0, z0, i, Nmax = 1000):
-    if z0 is not None:
-        ax.axvline(z0[i], color='r', ls=':')
-    z, p = re.posterior(x0, i, Nmax = Nmax)
-    ax.plot(z, p, 'k')
+def comb2d(indices):
+    output = []
+    for i in range(len(indices)):
+        for j in range(i+1, len(indices)):
+            output.append([indices[i], indices[j]])
+    return output
 
-def plot1d(re1d, x0, dims = (15, 5), ncol = 6, params = None, labels = None, z0 = None, cmap = 'Greys', Nmax = 1000):
-    # TODO: Rewrite
-    if params is None:
-        params = range(re1d.zdim)
+def combine_z(z, combinations):
+    """Generate parameter combinations in last dimension. 
+    Requires: z.ndim == 1. 
+    output.shape == (n_posteriors, parameter shape)
+    """
+    if combinations is None:
+        return z.unsqueeze(-1)
+    else:
+        return torch.stack([z[c] for c in combinations])
 
-    K = len(params)
-    nrow = (K-1)//ncol+1
+def set_device(gpu: bool = False) -> torch.device:
+    if gpu and torch.cuda.is_available():
+        device = torch.device("cuda")
+        torch.set_default_tensor_type("torch.cuda.FloatTensor")
+    elif gpu and not torch.cuda.is_available():
+        warn("Although the gpu flag was true, the gpu is not avaliable.")
+        device = torch.device("cpu")
+        torch.set_default_tensor_type("torch.FloatTensor")
+    else:
+        device = torch.device("cpu")
+        torch.set_default_tensor_type("torch.FloatTensor")
+    return device
 
-    fig, axes = plt.subplots(nrow, ncol, figsize=dims)
-    lb = 0.125
-    tr = 0.9
-    whspace = 0.15
-    fig.subplots_adjust(left=lb, bottom=lb, right=tr, top=tr, wspace=whspace, hspace=whspace)
-    
-    if labels is None:
-        labels = ['z%i'%params[i] for i in range(K)]
-    for k in range(K):
-        if nrow == 1:
-            ax = axes[k]
-        else:
-            i, j = k%ncol, k//ncol
-            ax = axes[j, i]
-        hist1d(ax, re1d, x0, z0, params[k], Nmax = Nmax)
-        ax.set_xlabel(labels[k])
+def get_x(list_xz):
+    """Extract x from batch of samples."""
+    return [xz['x'] for xz in list_xz]
 
-def corner(re1d, re2d, x0, dim = 10, params = None, labels = None, z0 = None, cmap = 'Greys', Nmax = 1000):
-    # TODO: Rewrite
-    if params is None:
-        params = range(re1d.zdim)
+def get_z(list_xz):
+    """Extract z from batch of samples."""
+    return [xz['z'] for xz in list_xz]
 
-    K = len(params)
-    fig, axes = plt.subplots(K, K, figsize=(dim, dim))
-    lb = 0.125
-    tr = 0.9
-    whspace = 0.1
-    fig.subplots_adjust(left=lb, bottom=lb, right=tr, top=tr, wspace=whspace, hspace=whspace)
-    
-    if labels is None:
-        labels = ['z%i'%params[i] for i in range(K)]
-    for i in range(K):
-        for j in range(K):
-            ax = axes[i, j]
-            # Switch off upper left triangle
-            if i < j:
-                ax.set_yticklabels([])
-                ax.set_xticklabels([])
-                ax.set_xticks([])
-                ax.set_yticks([])
-                ax.set_frame_on(False)
-                continue
 
-            # Formatting labels
-            if j > 0 or i == 0:
-                ax.set_yticklabels([])
-                #ax.set_yticks([])
-            if i < K-1:
-                ax.set_xticklabels([])
-                #ax.set_xticks([])
-            if i == K-1:
-                ax.set_xlabel(labels[j])
-            if j == 0 and i > 0:
-                ax.set_ylabel(labels[i])
-
-            # Set limits
-            #ax.set_xlim(x_lims[j])
-            #if i != j:
-            #    ax.set_ylim(y_lims[i])
-
-            # 2-dim plots
-            if j < i:
-                cont2d(ax, re2d, x0, z0, params[j], params[i], cmap = cmap, Nmax = Nmax)
-
-            if j == i:
-                hist1d(ax, re1d, x0, z0, params[i], Nmax = Nmax)
+if __name__ == "__main__":
+    pass
